services/spaces_service/requirements.txt
django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.4.0
django-environ==0.11.2
drf-spectacular==0.27.2
PyMySQL==1.1.1
gunicorn==21.2.0
requests==2.32.3
dj-database-url==2.1.0
whitenoise==6.6.0
